Make Galera service A/P
The service created for a Galera cluster always balances traffic
to all available Galera replicas, which means the cluster is
effectively configured as an A/A service.

In order to limit the burden on clients of supporting A/A semantics,
change the way the service is configured to act as A/P instead.

A new callback script is introduced in pods. It is called by mysqld
automatically every time the Galera library detects a change in the
cluster (node join, crash, network partition...). This callback
script reconfigures the service CR via label selectors to drive
traffic to a single pod at a time.

Jira: OSPRH-7405